requirements.txt
moviepy==1.0.3
Pillow==8.2.0
progress==1.5
numpy==1.20.3
opencv-python==4.5.2.54